+++ event/src/main/java/com/devticket/event/common/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.devticket.event.infrastructure.config;
+package com.devticket.event.common.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
