+++ event/src/main/java/com/devticket/event/common/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.devticket.event.domain.exception;
+package com.devticket.event.common.exception;
 
 public interface ErrorCode {
     int getStatus();
